Fix issue with root user caching for API commands
When we create a `botUser`, we first create a root user and log
the root user in as the bot. The `withCache` option of the `root`
function is used in `botUser` to ensure that we get a fresh `root`
user every time to log in with.

Unforunately, the `root` function was also writing the created
user object to the state cache, even if `withCache` was set to
false. This meant that after we create a `botUser`, we were saving
a root-user-logged-in-as-bot to the root user slot in the cache.

Change the `root` function so that we also don't write the newly-
created user to the cache if `withCache` is set to false.

Bug: T372140